pyproject.toml
[project]
name = "notetrans"
version = "0.1.0"
description = "HackMD to Obsidian Markdown exporter"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "requests>=2.31",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.scripts]
notetrans = "notetrans.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "responses>=0.25",
]
